Hack around BTN3 and BTN2 being swapped
Hack: For some reason not yet understood, BTN3 shows up as Button 2 in
xev, and BTN2 shows up as Button 3 in xev. Swap them around in the
keymap. The physical order on the keyboard should be left-to-right,
just like it is on standard 3-button mice, to make it easier to
remember.

This probelm may exist only on my Ubuntu 16.04 LTS machine. Have to
check this in to solicit some help.