Add Qwerty layout and selector layout
The Fn key was one I never use. Repurpose it to switch between qwerty
(via Fn+q) or to the default layout (via Fn+d).  This was helpful to
switch to the qwerty layout because certain games treat the "a" key as
an arrow key used by the left hand, and we need to be able to hold it
down to repeat.